+++ src/__tests__/calculatorpage.test.js
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom';
 import React from 'react';
 import { render } from '@testing-library/react';
-import CalculatorPage from './components/CalculatorPage';
+import CalculatorPage from '../components/CalculatorPage';
 
 it('test calculatorpage', () => {
   render(<CalculatorPage />);
