+++ src/__tests__/home.test.js
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom';
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Home from './components/Home';
+import Home from '../components/Home';
 
 test('To test for display', () => {
   render(<Home />);
