+++ src/__tests__/title.test.js
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom';
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Title from './components/title';
+import Title from '../components/title';
 
 test('test title.js', () => {
   render(<Title />);
